refactor product store function

+++ Modules/Product/Http/Controllers/ProductController.php
@@ -2,6 +2,7 @@
 
 namespace Modules\Product\Http\Controllers;
 
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Contracts\View\Factory;
 use Illuminate\Contracts\View\View;
@@ -9,6 +10,7 @@ use Illuminate\Http\JsonResponse;
 use Illuminate\Http\RedirectResponse;
 use Modules\Category\Repositories\CategoryRepoEloquentInterface;
 use Modules\Product\Http\Requests\ProductRequest;
+use Modules\Product\Models\Product;
 use Modules\Product\Repositories\ProductRepoEloquentInterface;
 use Modules\Product\Services\ProductService;
 use Modules\Share\Http\Controllers\Controller;
@@ -17,6 +19,13 @@ use Modules\Share\Services\ShareService;
 
 class ProductController extends Controller
 {
+    /**
+     * Get product class.
+     *
+     * @var string
+     */
+    private string $class = Product::class;
+
     public ProductService $service;
     public ProductRepoEloquentInterface $repo;
 
@@ -30,9 +39,11 @@ class ProductController extends Controller
      * Show index page of products.
      *
      * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View
+     * @throws AuthorizationException
      */
     public function index()
     {
+        $this->authorize('manage', $this->class);
         $products = $this->repo->getLatest()
             ->with(['categories', 'first_media', 'second_media', 'tags'])
             ->paginate();
@@ -44,9 +55,11 @@ class ProductController extends Controller
      * Show create product page.
      *
      * @return Application|Factory|View
+     * @throws AuthorizationException
      */
     public function create(CategoryRepoEloquentInterface $categoryRepo)
     {
+        $this->authorize('manage', $this->class);
         $categories = $categoryRepo->getActiveCategories()->get();
 
         return view('Product::create', compact('categories'));
@@ -58,29 +71,20 @@ class ProductController extends Controller
      * @param  ProductRequest $request
      * @return RedirectResponse
      * @throws \Exception
+     * @throws AuthorizationException
      */
     public function store(ProductRequest $request)
     {
-        ShareService::uploadMediaWithAddInRequest($request, 'first_media', 'first_media_id');
-        ShareService::uploadMediaWithAddInRequest($request, 'second_media', 'second_media_id');
+        $this->authorize('manage', $this->class);
+        $this->uploadMedias($request);
 
-        // Convert
-        $request->price = str_replace(',', '', $request->price);
-
-        $product = $this->service->store($request);
+        $product = $this->service->store($request->all());
 
         $this->service->attachCategoriesToProduct($request->categories, $product);
         $this->service->attachGalleriesToProduct($request->galleries, $product);
 
-        $attributes = $request->input('attributes');
-        if (! is_null($attributes[0]['attributekeys'])) {
-            $this->service->attachAttributesToProduct($attributes, $product);
-        }
-
-        $tags = $request->tags;
-        if (! empty($tags)) {
-            $this->service->attachTagsToProduct($tags, $product);
-        }
+        $this->checkAndAttachAttributes($request->input('attributes'), $product);
+        $this->checkAndAttachTags($request->tags, $product);
 
         return $this->successMessageWithRedirect('Create product');
     }
@@ -91,9 +95,12 @@ class ProductController extends Controller
      * @param  $id
      * @param  CategoryRepoEloquentInterface $categoryRepo
      * @return Application|Factory|View
+     * @throws AuthorizationException
      */
     public function edit($id, CategoryRepoEloquentInterface $categoryRepo)
     {
+        $this->authorize('manage', $this->class);
+
         $product = $this->repo->findById($id);
         $categories = $categoryRepo->getActiveCategories()->get();
 
@@ -106,13 +113,12 @@ class ProductController extends Controller
      * @param  ProductRequest $request
      * @param  $id
      * @return RedirectResponse
+     * @throws AuthorizationException
      */
     public function update(ProductRequest $request, $id)
     {
+        $this->authorize('manage', $this->class);
         $product = $this->repo->findById($id);
-
-        // Convert
-        $request->price = str_replace(',', '', $request->price);
 
         if (! is_null($request->first_media)) {
             ShareService::uploadMediaWithAddInRequest($request, 'first_media', 'first_media_id');
@@ -141,9 +147,11 @@ class ProductController extends Controller
      *
      * @param  $id
      * @return JsonResponse
+     * @throws AuthorizationException
      */
     public function destroy($id)
     {
+        $this->authorize('manage', $this->class);
         $this->repo->delete($id);
 
         return AjaxResponses::SuccessResponse();
@@ -159,5 +167,25 @@ class ProductController extends Controller
     {
         ShareService::successToast($title);
         return to_route('products.index');
+    }
+
+    private function checkAndAttachAttributes($attributes, $product): void
+    {
+        if (!is_null($attributes[0]['attributekeys'])) {
+            $this->service->attachAttributesToProduct($attributes, $product);
+        }
+    }
+
+    private function checkAndAttachTags($tags, $product): void
+    {
+        if (!empty($tags)) {
+            $this->service->attachTagsToProduct($tags, $product);
+        }
+    }
+
+    private function uploadMedias(ProductRequest $request): void
+    {
+        ShareService::uploadMediaWithAddInRequest($request, 'first_media', 'first_media_id');
+        ShareService::uploadMediaWithAddInRequest($request, 'second_media', 'second_media_id');
     }
 }
